backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
websockets>=12.0
numpy>=1.24
scipy>=1.11
librosa>=0.10
python-rtmidi>=1.5
